Moves JUnitFailureHacker to the main junit internal package
